+++ gems/aws-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased Changes
 ------------------
 
 * Issue - Perform a case-insensitive comparison when filtering sensitive parameters from logs
+* Issue - Add passthrough of region from client to STS when using assume_role_web_identity_credentials
 
 3.90.0 (2020-02-12)
 ------------------